+++ modular_robot_simulation/revolve2/modular_robot_simulation/_build_multi_body_systems/_builders/_attachment_face_builder.py
@@ -1,22 +1,19 @@
-from revolve2.modular_robot.body.base import Core
-from revolve2.simulation.scene import AABB, MultiBodySystem, Pose, RigidBody
-from revolve2.simulation.scene.geometry import GeometryBox
-from revolve2.simulation.scene.geometry.textures import Texture
+from revolve2.modular_robot.body.base import AttachmentFace
+from revolve2.simulation.scene import MultiBodySystem, Pose, RigidBody
 
-from ._body_to_multi_body_system_mapping import BodyToMultiBodySystemMapping
+from .._body_to_multi_body_system_mapping import BodyToMultiBodySystemMapping
+from .._unbuilt_child import UnbuiltChild
 from ._builder import Builder
-from ._convert_color import convert_color
-from ._unbuilt_child import UnbuiltChild
 
 
-class CoreBuilder(Builder):
-    """A Builder for Cores."""
+class AttachmentFaceBuilder(Builder):
+    """A Builder for Attachment Faces."""
 
-    _module: Core
+    _module: AttachmentFace
 
-    def __init__(self, module: Core, rigid_body: RigidBody, slot_pose: Pose):
+    def __init__(self, module: AttachmentFace, rigid_body: RigidBody, slot_pose: Pose):
         """
-        Initialize the Core V1 Builder.
+        Initialize the Attachment Face Builder.
 
         :param module: The module to be built.
         :param rigid_body: The rigid body for the module to be built on.
@@ -38,15 +35,6 @@ class CoreBuilder(Builder):
         :param body_to_multi_body_system_mapping: A mapping from body to multi-body system
         :return: The next children to be built.
         """
-        self._rigid_body.geometries.append(
-            GeometryBox(
-                pose=self._slot_pose,
-                mass=self._module.mass,
-                texture=Texture(base_color=convert_color(self._module.color)),
-                aabb=AABB(self._module.bounding_box),
-            )
-        )
-
         tasks = []
         for child_index, attachment_point in self._module.attachment_points.items():
             child = self._module.children.get(child_index)
@@ -57,12 +45,8 @@ class CoreBuilder(Builder):
                 )
                 unbuilt.make_pose(
                     position=self._slot_pose.position
-                    + self._slot_pose.orientation
-                    * attachment_point.orientation
-                    * attachment_point.offset,
-                    orientation=self._slot_pose.orientation
-                    * attachment_point.orientation,
+                    + self._slot_pose.orientation * attachment_point.offset,
+                    orientation=self._slot_pose.orientation,
                 )
-
                 tasks.append(unbuilt)
         return tasks
