+++ modular_robot_simulation/revolve2/modular_robot_simulation/_build_multi_body_systems/_builders/_brick_builder.py
@@ -5,10 +5,10 @@ from revolve2.simulation.scene import AABB, MultiBodySystem, Pose, RigidBody
 from revolve2.simulation.scene.geometry import GeometryBox
 from revolve2.simulation.scene.geometry.textures import Texture
 
-from ._body_to_multi_body_system_mapping import BodyToMultiBodySystemMapping
+from .._body_to_multi_body_system_mapping import BodyToMultiBodySystemMapping
+from .._convert_color import convert_color
+from .._unbuilt_child import UnbuiltChild
 from ._builder import Builder
-from ._convert_color import convert_color
-from ._unbuilt_child import UnbuiltChild
 
 
 class BrickBuilder(Builder):
