+++ modular_robot_simulation/revolve2/modular_robot_simulation/_build_multi_body_systems/_builders/_builder.py
@@ -2,8 +2,8 @@ from abc import ABC, abstractmethod
 
 from revolve2.simulation.scene import MultiBodySystem
 
-from ._body_to_multi_body_system_mapping import BodyToMultiBodySystemMapping
-from ._unbuilt_child import UnbuiltChild
+from .._body_to_multi_body_system_mapping import BodyToMultiBodySystemMapping
+from .._unbuilt_child import UnbuiltChild
 
 
 class Builder(ABC):
