+++ modular_robot_simulation/revolve2/modular_robot_simulation/_build_multi_body_systems/_get_builder.py
@@ -1,10 +1,12 @@
 from revolve2.modular_robot.body.base import ActiveHinge, AttachmentFace, Brick, Core
 
-from ._active_hinge_builder import ActiveHingeBuilder
-from ._attachment_face_builder import AttachmentFaceBuilder
-from ._brick_builder import BrickBuilder
-from ._builder import Builder
-from ._core_builder import CoreBuilder
+from ._builders import (
+    ActiveHingeBuilder,
+    AttachmentFaceBuilder,
+    BrickBuilder,
+    Builder,
+    CoreBuilder,
+)
 from ._unbuilt_child import UnbuiltChild
